requirements.txt
dvc
mlxtend
torch
torchvision
PyYAML
numpy
pandas
scikit-learn
matplotlib
jupyter
notebook
